24-10-06

+++ src/main/java/org/example/edufyalbumforartist/repostitories/AlbumRepository.java
@@ -3,5 +3,5 @@ package org.example.edufyalbumforartist.repostitories;
 import org.example.edufyalbumforartist.entities.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlbumRepository extends JpaRepository<Album, Integer> {
+public interface AlbumRepository extends JpaRepository<Album, Long> {
 }
