+++ src/main/java/org/example/edufyalbumforartist/services/ArtistService.java
@@ -20,7 +20,7 @@ public class ArtistService implements ArtistServiceInterface{
     @Autowired
     private AlbumRepository albumRepository;
 
-    public Set<Album> getAlbumsByArtist(Integer artistId) {
+    public Set<Album> getAlbumsByArtist(Long artistId) {
         Artist artist = artistRepository.findById(artistId)
                 .orElseThrow(() -> new ResourceNotFoundException("Artist", "ID", artistId));
         return artist.getAlbums();
@@ -34,13 +34,13 @@ public class ArtistService implements ArtistServiceInterface{
         return artistRepository.findAll();
     }
 
-    public void deleteArtistById(Integer artistId) {
+    public void deleteArtistById(Long artistId) {
         artistRepository.findById(artistId).
                 orElseThrow(() -> new ResourceNotFoundException("Artist", "ID", artistId));
         artistRepository.deleteById(artistId);
     }
 
-    public Artist updateArtist(Integer artistId, Artist artist) {
+    public Artist updateArtist(Long artistId, Artist artist) {
         Artist updatedArtist = artistRepository.findById(artistId)
                 .orElseThrow(() -> new ResourceNotFoundException("Artist", "ID", artistId));
         if (artist.getName() != null) {
